+++ Sources/QuickJSON/Decoding.swift
@@ -9,6 +9,34 @@ import Logging
 #if QUICKJSON_SHOULDLOG
 /// decode an unknown type from a json document using a specified handler function.
 /// - parameters:
+/// 	- bytes: the json document to decode
+/// 	- flags: the decoding flags to use
+/// 	- memconfig: the memory configuration to use
+/// 	- logLevel: the log level to use for this job
+/// 	- handlerFunc: the function to handle the parsing actions
+/// - returns: transparently returns the return value of the handler function
+public func decode<T:Decodable, R, C>(
+	bytes:C,
+	flags:Decoding.Flags = Decoding.Flags(),
+	memory memconfig:Memory.Configuration = .automatic,
+	logLevel:Logging.Logger.Level = .critical,
+	_ handlerFunc:(Swift.Decoder) throws -> R
+) throws -> R where C:Collection, C.Element == UInt8 {
+	let getVal = try bytes.withContiguousStorageIfAvailable({
+		return try decode(data:$0.baseAddress!, size:$0.count, flags:flags, memory:memconfig, logLevel:logLevel, handlerFunc)
+	})
+	if getVal != nil {
+		return getVal!
+	} else {
+		let buffer = UnsafeMutableBufferPointer<UInt8>.allocate(capacity: self.count)
+		defer { buffer.deallocate() }
+		_ = buffer.initialize(from: self)
+		return try decode(data:buffer.baseAddress!, size:buffer.count, flags:flags, memory:memconfig, logLevel:logLevel, handlerFunc)
+	}
+}
+
+/// decode an unknown type from a json document using a specified handler function.
+/// - parameters:
 /// 	- data: a pointer to the json document to decode
 /// 	- size: the size of the json document
 /// 	- flags: the decoding flags to use
@@ -54,6 +82,33 @@ public func decode<T:Decodable, R>(
 /// 	- memconfig: the memory configuration to use
 /// 	- handlerFunc: the function to handle the parsing actions
 /// - returns: transparently returns the return value of the handler function
+public func decode<R, C>(
+	bytes:C,
+	flags:Decoding.Flags = Decoding.Flags(),
+	memory memconfig:Memory.Configuration = .automatic,
+	_ handlerFunc:(Swift.Decoder) throws -> R
+) throws -> R where C:Collection, C.Element == UInt8 {
+	let getVal = try bytes.withContiguousStorageIfAvailable({
+		return try decode(data:$0.baseAddress!, size:$0.count, flags:flags, memory:memconfig, handlerFunc)
+	})
+	if getVal != nil {
+		return getVal!
+	} else {
+		let buffer = UnsafeMutableBufferPointer<UInt8>.allocate(capacity:bytes.count)
+		defer { buffer.deallocate() }
+		_ = buffer.initialize(from: bytes)
+		return try decode(data:buffer.baseAddress!, size:buffer.count, flags:flags, memory:memconfig, handlerFunc)
+	}
+}
+
+/// decode an unknown type from a json document using a specified handler function.
+/// - parameters:
+/// 	- data: a pointer to the json document to decode
+/// 	- size: the size of the json document
+/// 	- flags: the decoding flags to use
+/// 	- memconfig: the memory configuration to use
+/// 	- handlerFunc: the function to handle the parsing actions
+/// - returns: transparently returns the return value of the handler function
 public func decode<R>(
 	data:UnsafeRawPointer, size:size_t,
 	flags:Decoding.Flags = Decoding.Flags(),
@@ -86,9 +141,42 @@ public func decode<R>(
 
 #if QUICKJSON_SHOULDLOG
 /// decode a value from a json document
-/// - parameter type: the type of the value to decode
-/// - parameter data: the pointer to the json document to decode
-/// - parameter flags: decoding option flags
+///	- parameters:
+///		- type: the type of the value to decode
+///		- bytes: the json document to decode
+///		- flags: the decoding flags to use
+///		- memconfig: the memory configuration to use
+///		- logLevel: the log level to use for this job
+///	- returns: the decoded value
+public func decode<T:Decodable, C>(
+	_ type:T.Type, 
+	bytes:C,
+	flags:Decoding.Flags = Decoding.Flags(), 
+	memory memconfig:Memory.Configuration = .automatic, 
+	logLevel:Logging.Logger.Level = .critical
+) throws -> T where C:Collection, C.Element == UInt8 {
+	let getVal = try bytes.withContiguousStorageIfAvailable({
+		return try decode(type, data:$0.baseAddress!, size:$0.count, flags:flags, memory:memconfig, logLevel:logLevel)
+	})
+	if getVal != nil {
+		return getVal!
+	} else {
+		let buffer = UnsafeMutableBufferPointer<UInt8>.allocate(capacity:bytes.count)
+		defer { buffer.deallocate() }
+		_ = buffer.initialize(from: bytes)
+		return try decode(type, data:buffer.baseAddress!, size:buffer.count, flags:flags, memory:memconfig, logLevel:logLevel)
+	}
+}
+
+/// decode a value from a json document
+///	- parameters:
+///		- type: the type of the value to decode
+///		- data: the pointer to the json document to decode
+///		- size: the size of the json document buffer
+///		- flags: the decoding flags to use
+///		- memconfig: the memory configuration to use
+///		- logLevel: the log level to use for this job
+///	- returns: the decoded value
 public func decode<T:Decodable>(
 	_ type:T.Type, 
 	from data:UnsafeRawPointer, size:size_t, 
@@ -119,10 +207,43 @@ public func decode<T:Decodable>(
 	return try T(from:decoder(root:getRoot!, logLevel:logLevel))
 }
 #else
+
 /// decode a value from a json document
-/// - parameter type: the type of the value to decode
-/// - parameter data: the pointer to the json document to decode
-/// - parameter flags: decoding option flags
+/// - parameters:
+///		- type: the type of the value to decode
+///		- bytes: the json document to decode
+///		- size: the size of the json document buffer
+///		- flags: the decoding flags to use
+///		- memconfig: the memory configuration to use
+///	- returns: the decoded value
+public func decode<T:Decodable, C>(
+	_ type:T.Type, 
+	bytes:C, 
+	size:size_t, 
+	flags:Decoding.Flags = Decoding.Flags(), 
+	memory memconfig:Memory.Configuration = .automatic
+) throws -> T where C:Collection, C.Element == UInt8 {
+	let getVal = try bytes.withContiguousStorageIfAvailable({
+		return try decode(type, from:$0.baseAddress!, size:$0.count, flags:flags, memory:memconfig)
+	})
+	if getVal != nil {
+		return getVal!
+	} else {
+		let buffer = UnsafeMutableBufferPointer<UInt8>.allocate(capacity:bytes.count)
+		defer { buffer.deallocate() }
+		_ = buffer.initialize(from: bytes)
+		return try decode(type, from:buffer.baseAddress!, size:buffer.count, flags:flags, memory:memconfig)
+	}
+}
+
+/// decode a value from a json document
+///	- parameters:
+///		- type: the type of the value to decode
+///		- data: the pointer to the json document to decode
+///		- size: the size of the json document buffer
+///		- flags: the decoding flags to use
+///		- memconfig: the memory configuration to use
+///	- returns: the decoded value
 public func decode<T:Decodable>(
 	_ type:T.Type, 
 	from data:UnsafeRawPointer, 
@@ -153,6 +274,7 @@ public func decode<T:Decodable>(
 	return try T(from:decoder(root:getRoot!))
 }
 #endif
+
 /// namespace related to decoding.
 public struct Decoding {
 	/// errors that can be thrown by the decoder